% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_pen.R
\name{calc_pen}
\alias{calc_pen}
\title{Extrapolation of precipitation depths for Tn > 100 a according to PEN-LAWA}
\usage{
calc_pen(x = NULL)
}
\arguments{
\item{x}{Tibble containing grid cell statistics from KOSTRA-DWD-2010R,
as provided by \code{get_stats()}.}
}
\value{
Tibble containing extrapolated precipitation depths as a function of
duration and return periods.
}
\description{
Extrapolation of precipitation depths for Tn > 100 a according to PEN-LAWA
}
\examples{
stats <- get_stats("42016")

calc_pen(stats)
}
\references{
Verworn & Kummer (2006): Praxisrelevante Extremwerte des Niederschlages (PEN).\cr
Verworn & Draschoff (2008): PEN-Erweiterung.
}
\seealso{
\code{\link[=get_stats]{get_stats()}}
}
