% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_depth.R
\name{get_depth}
\alias{get_depth}
\title{Get precipitation depth for a specified duration level and return period}
\usage{
get_depth(x = NULL, d = NULL, tn = NULL, uc = FALSE)
}
\arguments{
\item{x}{Tibble containing grid cell statistics from KOSTRA-DWD-2010R,
as provided by \code{get_stats()}.}

\item{d}{numeric. Precipitation duration level \code{[min]}.}

\item{tn}{numeric. Return period \code{[a]}.}

\item{uc}{logical. Consider overall uncertainties,
as provided by \code{get_uncertainties()}?}
}
\value{
units. Precipitation depth \code{[mm]}.
}
\description{
Get precipitation depth for a specified duration level and return period
}
\examples{
stats <- get_stats("42016")

get_depth(stats, d = 60, tn = 50)
get_depth(stats, d = 60, tn = 50, uc = TRUE)
}
\seealso{
\code{\link[=get_stats]{get_stats()}}, \code{\link[=get_uncertainties]{get_uncertainties()}}
}
