% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_spatial.R
\name{view_spatial}
\alias{view_spatial}
\title{Interactive visualisation of tile locations from KOSTRA-DWD-2010R dataset}
\usage{
view_spatial(x = NULL, file = NULL)
}
\arguments{
\item{x}{character. Relevant "INDEX_RC" field to be queried.}

\item{file}{(optional) character. Filename with png extension to be used for data output.}
}
\value{
Object of class \code{c("leaflet", "htmlwidget")}.
}
\description{
Interactive visualisation of tile locations from KOSTRA-DWD-2010R dataset
}
\examples{
\dontrun{
view_spatial()

view_spatial("49011")
view_spatial("49011", file = "kostra2010R_49011.png")
}
}
